+++ src/metropolis.h
@@ -1,8 +1,8 @@
 #ifndef METROPOLIS_H
 #define METROPOLIS_H
 
-int metropolis(int * lattice, int n, float T);
+int metropolis(int * lattice, int n, double T);
 int pick_site(int * lattice, int n);
-int flip(int * lattice, int n, float T, int site);
+int flip(int * lattice, int n, double T, int site);
 
 #endif
